+++ docs/rules/no-nonstandard-arraybuffer-properties.md
@@ -1,9 +1,9 @@
 ---
-title: "es-x/no-nonstandard-array-buffer-properties"
+title: "es-x/no-nonstandard-arraybuffer-properties"
 description: "disallow non-standard properties on `ArrayBuffer` class"
 ---
 
-# es-x/no-nonstandard-array-buffer-properties
+# es-x/no-nonstandard-arraybuffer-properties
 > disallow non-standard properties on `ArrayBuffer` class
 
 - ❗ <badge text="This rule has not been released yet." vertical="middle" type="error"> ***This rule has not been released yet.*** </badge>
@@ -17,7 +17,7 @@ This rule reports non-standard properties on `ArrayBuffer` class as errors.
 <eslint-playground type="bad">
 
 ```js
-/*eslint es-x/no-nonstandard-array-buffer-properties: error */
+/*eslint es-x/no-nonstandard-arraybuffer-properties: error */
 ArrayBuffer.unknown();
 ```
 
@@ -30,7 +30,7 @@ This rule has an option.
 ```jsonc
 {
   "rules": {
-    "es-x/no-nonstandard-array-buffer-properties": [
+    "es-x/no-nonstandard-arraybuffer-properties": [
       "error",
       {
         "allow": []
@@ -46,5 +46,5 @@ An array of non-standard property names to allow.
 
 ## 📚 References
 
-- [Rule source](https://github.com/eslint-community/eslint-plugin-es-x/blob/master/lib/rules/no-nonstandard-array-buffer-properties.js)
-- [Test source](https://github.com/eslint-community/eslint-plugin-es-x/blob/master/tests/lib/rules/no-nonstandard-array-buffer-properties.js)
+- [Rule source](https://github.com/eslint-community/eslint-plugin-es-x/blob/master/lib/rules/no-nonstandard-arraybuffer-properties.js)
+- [Test source](https://github.com/eslint-community/eslint-plugin-es-x/blob/master/tests/lib/rules/no-nonstandard-arraybuffer-properties.js)
