+++ lib/index.js
@@ -265,10 +265,10 @@ module.exports = {
         "no-math-trunc": require("./rules/no-math-trunc"),
         "no-modules": require("./rules/no-modules"),
         "no-new-target": require("./rules/no-new-target"),
-        "no-nonstandard-array-buffer-properties": require("./rules/no-nonstandard-array-buffer-properties"),
-        "no-nonstandard-array-buffer-prototype-properties": require("./rules/no-nonstandard-array-buffer-prototype-properties"),
         "no-nonstandard-array-properties": require("./rules/no-nonstandard-array-properties"),
         "no-nonstandard-array-prototype-properties": require("./rules/no-nonstandard-array-prototype-properties"),
+        "no-nonstandard-arraybuffer-properties": require("./rules/no-nonstandard-arraybuffer-properties"),
+        "no-nonstandard-arraybuffer-prototype-properties": require("./rules/no-nonstandard-arraybuffer-prototype-properties"),
         "no-nonstandard-atomics-properties": require("./rules/no-nonstandard-atomics-properties"),
         "no-nonstandard-bigint-properties": require("./rules/no-nonstandard-bigint-properties"),
         "no-nonstandard-bigint-prototype-properties": require("./rules/no-nonstandard-bigint-prototype-properties"),
@@ -312,16 +312,16 @@ module.exports = {
         "no-nonstandard-regexp-properties": require("./rules/no-nonstandard-regexp-properties"),
         "no-nonstandard-regexp-prototype-properties": require("./rules/no-nonstandard-regexp-prototype-properties"),
         "no-nonstandard-set-prototype-properties": require("./rules/no-nonstandard-set-prototype-properties"),
-        "no-nonstandard-shared-array-buffer-prototype-properties": require("./rules/no-nonstandard-shared-array-buffer-prototype-properties"),
+        "no-nonstandard-sharedarraybuffer-prototype-properties": require("./rules/no-nonstandard-sharedarraybuffer-prototype-properties"),
         "no-nonstandard-string-properties": require("./rules/no-nonstandard-string-properties"),
         "no-nonstandard-string-prototype-properties": require("./rules/no-nonstandard-string-prototype-properties"),
         "no-nonstandard-symbol-properties": require("./rules/no-nonstandard-symbol-properties"),
         "no-nonstandard-symbol-prototype-properties": require("./rules/no-nonstandard-symbol-prototype-properties"),
         "no-nonstandard-typed-array-properties": require("./rules/no-nonstandard-typed-array-properties"),
         "no-nonstandard-typed-array-prototype-properties": require("./rules/no-nonstandard-typed-array-prototype-properties"),
-        "no-nonstandard-weak-map-prototype-properties": require("./rules/no-nonstandard-weak-map-prototype-properties"),
-        "no-nonstandard-weak-ref-prototype-properties": require("./rules/no-nonstandard-weak-ref-prototype-properties"),
-        "no-nonstandard-weak-set-prototype-properties": require("./rules/no-nonstandard-weak-set-prototype-properties"),
+        "no-nonstandard-weakmap-prototype-properties": require("./rules/no-nonstandard-weakmap-prototype-properties"),
+        "no-nonstandard-weakref-prototype-properties": require("./rules/no-nonstandard-weakref-prototype-properties"),
+        "no-nonstandard-weakset-prototype-properties": require("./rules/no-nonstandard-weakset-prototype-properties"),
         "no-nullish-coalescing-operators": require("./rules/no-nullish-coalescing-operators"),
         "no-number-epsilon": require("./rules/no-number-epsilon"),
         "no-number-isfinite": require("./rules/no-number-isfinite"),
