+++ lib/rules/no-nonstandard-arraybuffer-properties.js
@@ -1,20 +1,18 @@
 "use strict"
 
 const {
-    defineNonstandardPrototypePropertiesHandler,
-} = require("../util/define-nonstandard-prototype-properties-handler")
-const {
-    arrayBufferPrototypeProperties,
-} = require("../util/well-known-properties")
+    defineNonstandardPropertiesHandler,
+} = require("../util/define-nonstandard-properties-handler")
+const { arrayBufferProperties } = require("../util/well-known-properties")
 
 module.exports = {
     meta: {
         docs: {
             description:
-                "disallow non-standard properties on ArrayBuffer instance",
+                "disallow non-standard properties on `ArrayBuffer` class",
             category: "nonstandard",
             recommended: false,
-            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-array-buffer-prototype-properties.html",
+            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-arraybuffer-properties.html",
         },
         fixable: null,
         messages: {
@@ -39,9 +37,9 @@ module.exports = {
         /** @type {Set<string>} */
         const allows = new Set([
             ...(context.options[0]?.allow || []),
-            ...arrayBufferPrototypeProperties,
+            ...arrayBufferProperties,
         ])
-        return defineNonstandardPrototypePropertiesHandler(
+        return defineNonstandardPropertiesHandler(
             context,
             "ArrayBuffer",
             allows,
