+++ lib/rules/no-nonstandard-arraybuffer-prototype-properties.js
@@ -1,18 +1,20 @@
 "use strict"
 
 const {
-    defineNonstandardPropertiesHandler,
-} = require("../util/define-nonstandard-properties-handler")
-const { arrayBufferProperties } = require("../util/well-known-properties")
+    defineNonstandardPrototypePropertiesHandler,
+} = require("../util/define-nonstandard-prototype-properties-handler")
+const {
+    arrayBufferPrototypeProperties,
+} = require("../util/well-known-properties")
 
 module.exports = {
     meta: {
         docs: {
             description:
-                "disallow non-standard properties on `ArrayBuffer` class",
+                "disallow non-standard properties on ArrayBuffer instance",
             category: "nonstandard",
             recommended: false,
-            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-array-buffer-properties.html",
+            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-arraybuffer-prototype-properties.html",
         },
         fixable: null,
         messages: {
@@ -37,9 +39,9 @@ module.exports = {
         /** @type {Set<string>} */
         const allows = new Set([
             ...(context.options[0]?.allow || []),
-            ...arrayBufferProperties,
+            ...arrayBufferPrototypeProperties,
         ])
-        return defineNonstandardPropertiesHandler(
+        return defineNonstandardPrototypePropertiesHandler(
             context,
             "ArrayBuffer",
             allows,
