+++ lib/rules/no-nonstandard-sharedarraybuffer-prototype-properties.js
@@ -3,15 +3,18 @@
 const {
     defineNonstandardPrototypePropertiesHandler,
 } = require("../util/define-nonstandard-prototype-properties-handler")
-const { weakRefPrototypeProperties } = require("../util/well-known-properties")
+const {
+    sharedArrayBufferPrototypeProperties,
+} = require("../util/well-known-properties")
 
 module.exports = {
     meta: {
         docs: {
-            description: "disallow non-standard properties on WeakRef instance",
+            description:
+                "disallow non-standard properties on SharedArrayBuffer instance",
             category: "nonstandard",
             recommended: false,
-            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-weak-ref-prototype-properties.html",
+            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-sharedarraybuffer-prototype-properties.html",
         },
         fixable: null,
         messages: {
@@ -36,11 +39,11 @@ module.exports = {
         /** @type {Set<string>} */
         const allows = new Set([
             ...(context.options[0]?.allow || []),
-            ...weakRefPrototypeProperties,
+            ...sharedArrayBufferPrototypeProperties,
         ])
         return defineNonstandardPrototypePropertiesHandler(
             context,
-            "WeakRef",
+            "SharedArrayBuffer",
             allows,
         )
     },
