+++ lib/rules/no-nonstandard-weakmap-prototype-properties.js
@@ -11,7 +11,7 @@ module.exports = {
             description: "disallow non-standard properties on WeakMap instance",
             category: "nonstandard",
             recommended: false,
-            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-weak-map-prototype-properties.html",
+            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-weakmap-prototype-properties.html",
         },
         fixable: null,
         messages: {
