+++ lib/rules/no-nonstandard-weakref-prototype-properties.js
@@ -3,18 +3,15 @@
 const {
     defineNonstandardPrototypePropertiesHandler,
 } = require("../util/define-nonstandard-prototype-properties-handler")
-const {
-    sharedArrayBufferPrototypeProperties,
-} = require("../util/well-known-properties")
+const { weakRefPrototypeProperties } = require("../util/well-known-properties")
 
 module.exports = {
     meta: {
         docs: {
-            description:
-                "disallow non-standard properties on SharedArrayBuffer instance",
+            description: "disallow non-standard properties on WeakRef instance",
             category: "nonstandard",
             recommended: false,
-            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-shared-array-buffer-prototype-properties.html",
+            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-weakref-prototype-properties.html",
         },
         fixable: null,
         messages: {
@@ -39,11 +36,11 @@ module.exports = {
         /** @type {Set<string>} */
         const allows = new Set([
             ...(context.options[0]?.allow || []),
-            ...sharedArrayBufferPrototypeProperties,
+            ...weakRefPrototypeProperties,
         ])
         return defineNonstandardPrototypePropertiesHandler(
             context,
-            "SharedArrayBuffer",
+            "WeakRef",
             allows,
         )
     },
