+++ lib/rules/no-nonstandard-weakset-prototype-properties.js
@@ -11,7 +11,7 @@ module.exports = {
             description: "disallow non-standard properties on WeakSet instance",
             category: "nonstandard",
             recommended: false,
-            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-weak-set-prototype-properties.html",
+            url: "http://eslint-community.github.io/eslint-plugin-es-x/rules/no-nonstandard-weakset-prototype-properties.html",
         },
         fixable: null,
         messages: {
